+++ lib/custom_widgets_toolkit.dart
@@ -9,10 +9,12 @@ export 'src/others/page_animation.dart' show TransitionType, PageAnimation, Cust
 // exports of custom widgets
 
 export 'src/widgets/custom_elevated_button.dart' show CustomElevatedButton;
-export 'src/widgets/custom_rich_text.dart'
-    show CustomRichText, CustomTextSpanData;
+export 'src/widgets/custom_rich_text.dart' show CustomRichText, CustomTextSpanData;
 export 'src/widgets/custom_text_button.dart' show CustomTextButton;
 export 'src/widgets/custom_text.dart' show CustomText;
 export 'src/widgets/custom_textfield.dart' show CustomTextfield;
-export 'src/widgets/loading_dialog.dart' show LoadingDialog, OrganicBackgroundEffect;
+export 'src/widgets/custom_dialog.dart' show CustomDialog;
 export 'src/widgets/custom_snackbar.dart' show CustomSnackBar, SnackBarVibe;
+
+// exports of extras
+export 'src/extras/organic_background_effect.dart' show OrganicBackgroundEffect;