+++ lib/src/others/custom_scroll_physics.dart
@@ -1,5 +1,6 @@
-import 'package:flutter/material.dart';
+
 import 'package:flutter/physics.dart';
+import 'package:flutter/widgets.dart';
 
 /// A custom [ScrollPhysics] that supports multiple behaviors:
 /// - Android (clamping)