+++ CHANGELOG.md
@@ -188,3 +188,9 @@
 ## 0.0.63
 - **Improved performance for LoadingDialog**
 - **Removed blur by default on LoadingDialog background**
+
+
+## 0.0.65
+- **Breaking Changes**
+- **Removed the dispose2 in CustomTextfield. Added an autoDispose toggle whether to automatically clear assigned Controller/FocusNode(Clears by default)**
+- **Replaced LoadingDialog with CustomDialog. Do not update if your project relies on LoadingDialog**